Redirect readers to the latest and greatest Megaparsec tutorial
I have no intention of keeping all these texts up to date when they are
virtually included in the latest Megaparsec tutorial. So I'll just keep that
one up to date.

Removing the other tutorials altogether is not a good idea because there may
be links to them on the Internet.